gitk: Alter the ordering for the "Tags and heads" view
In the "Tags and heads" view, the list of refs is globally sorted.
The list of local refs (heads) is separated by the remote refs.  This
change re-orders the view toi be: local refs, remote refs tracked by
local refs, remote refs, tags, and then other refs

In addition, a configuration option is added to enable or disable this
behavior.

Signed-off-by: Michael Rappazzo <rappazzo@gmail.com>